Change BitVector to use `size_t` instead of `uint32_t` to address
bits in the array.

This fixes #40.

+++ core/Lucy/Index/BitVecDelDocs.c
@@ -40,9 +40,9 @@ BitVecDelDocs_init(BitVecDelDocs *self, Folder *folder,
         RETHROW(error);
     }
     // Cast away const-ness of buffer as we have no immutable BitVector.
-    int32_t len    = (int32_t)InStream_Length(ivars->instream);
+    int64_t len    = InStream_Length(ivars->instream);
     ivars->bits    = (uint8_t*)InStream_Buf(ivars->instream, len);
-    ivars->cap     = (uint32_t)(len * 8);
+    ivars->cap     = (size_t)(len * 8);
     return self;
 }
